+++ rust/otel-arrow-rust/src/validation/collector.rs
@@ -15,22 +15,29 @@ use tokio::time::timeout;
 
 use tokio::sync::mpsc;
 
+use super::service_type::{start_test_receiver, ServiceType};
+
 const READY_TIMEOUT_SECONDS: u64 = 10;
 const READY_MESSAGE: &str = "Everything is ready.";
+pub const SHUTDOWN_TIMEOUT_SECONDS: u64 = 5;
+pub const RECEIVER_TIMEOUT_SECONDS: u64 = 10;
+pub const TEST_TIMEOUT_SECONDS: u64 = 20;
 
-pub static COLLECTOR_PATH: LazyLock<String> =
-    LazyLock::new(|| {
-        let default_path = "../../bin/otelarrowcol";
-        let path = std::env::var("OTEL_COLLECTOR_PATH").unwrap_or(default_path.to_string());
-        
-        // Check if the collector exists at the specified path
-        if !std::path::Path::new(&path).exists() {
-            eprintln!("Warning: OpenTelemetry collector not found at '{}'. Tests may fail.", path);
-            eprintln!("Set OTEL_COLLECTOR_PATH environment variable to the correct path or ensure the collector is built.");
-        }
-        
-        path
-    });
+pub static COLLECTOR_PATH: LazyLock<String> = LazyLock::new(|| {
+    let default_path = "../../bin/otelarrowcol";
+    let path = std::env::var("OTEL_COLLECTOR_PATH").unwrap_or(default_path.to_string());
+
+    // Check if the collector exists at the specified path
+    if !std::path::Path::new(&path).exists() {
+        eprintln!(
+            "Warning: OpenTelemetry collector not found at '{}'. Tests may fail.",
+            path
+        );
+        eprintln!("Set OTEL_COLLECTOR_PATH environment variable to the correct path or ensure the collector is built.");
+    }
+
+    path
+});
 
 /// TimeoutError represents an error when a receiver operation times out
 #[derive(Debug)]
@@ -80,7 +87,7 @@ where
         for line in buf_reader.lines() {
             if let Ok(line) = line {
                 eprintln!("[{}] {}", prefix, line);
-                
+
                 // If we need to check for a ready message
                 if let Some((ref tx, message)) = probe {
                     if line.contains(message) {
@@ -134,7 +141,8 @@ impl CollectorProcess {
         // Create a unique temporary config file for the collector with a random identifier
         // to prevent collision with other tests
         let random_id = format!("{:016x}", rand::random::<u64>());
-        let config_path = PathBuf::from(env::temp_dir()).join(format!("otel_collector_config_{}.yaml", random_id));
+        let config_path = PathBuf::from(env::temp_dir())
+            .join(format!("otel_collector_config_{}.yaml", random_id));
 
         // Write the config to the file
         let mut file = fs::File::create(&config_path)
@@ -174,7 +182,7 @@ impl CollectorProcess {
 
         // Now create a oneshot channel for the async side
         let (tokio_tx, tokio_rx) = tokio::sync::oneshot::channel();
-        
+
         // Spawn a thread to bridge between the sync and async worlds
         thread::spawn(move || {
             // Wait for the ready signal from the sync channel
@@ -196,7 +204,10 @@ impl CollectorProcess {
                 stderr_handle: Some(stderr_handle),
             }),
             Ok(Err(_)) => Err("Channel closed before receiving ready message".to_string()),
-            Err(_) => Err(format!("Timed out after waiting {:?} for collector to be ready", timeout_duration))
+            Err(_) => Err(format!(
+                "Timed out after waiting {:?} for collector to be ready",
+                timeout_duration
+            )),
         }
     }
 }
@@ -221,7 +232,13 @@ impl Drop for CollectorProcess {
 }
 
 /// Configuration generator
-pub fn generate_config(exporter_name: &str, receiver_name: &str, signal: &str, receiver_port: u16, exporter_port: u16) -> String {
+pub fn generate_config(
+    exporter_name: &str,
+    receiver_name: &str,
+    signal: &str,
+    receiver_port: u16,
+    exporter_port: u16,
+) -> String {
     format!(
         r#"
 receivers:
@@ -253,4 +270,103 @@ service:
       level: info
 "#
     )
+}
+
+/// TestContext contains all the necessary components for running a test
+pub struct TestContext<S: ServiceType> {
+    pub client: S::Client,
+    pub collector: CollectorProcess,
+    pub request_rx: TimeoutReceiver<S::Request>,
+    pub server_handle: tokio::task::JoinHandle<Result<(), tonic::transport::Error>>,
+}
+
+/// Generic test runner for telemetry signal tests
+///
+/// This function will:
+/// 1. Start a generic test receiver server
+/// 2. Start the OTel collector
+/// 3. Create a test context with client and receiver
+/// 4. Run the supplied test logic
+/// 5. Perform cleanup
+///
+/// The service type parameter S determines which signal type to test (traces, metrics, or logs)
+pub async fn run_test<S, T, F>(test_logic: F) -> Result<(), Box<dyn std::error::Error>>
+where
+    S: ServiceType,
+    S::Request: std::fmt::Debug + PartialEq,
+    F: FnOnce(TestContext<S>) -> T,
+    T: std::future::Future<Output = (TestContext<S>, Result<(), Box<dyn std::error::Error>>)>,
+{
+    // Generate random ports in the high u16 range to avoid conflicts
+    let random_value = rand::random::<u16>();
+    let receiver_port = 40000 + (random_value % 25000);
+
+    // Start the test receiver server and wrap it with a timeout to avoid tests getting stuck
+    let (server_handle, request_rx_raw, exporter_port) = start_test_receiver::<S>()
+        .await
+        .map_err(|e| format!("Failed to start test receiver: {}", e))?;
+
+    // Create a timeout-wrapped version of the receiver
+    let timeout_duration = std::time::Duration::from_secs(RECEIVER_TIMEOUT_SECONDS);
+    let request_rx = TimeoutReceiver {
+        inner: request_rx_raw,
+        timeout: timeout_duration,
+    };
+
+    // Generate and start the collector with OTLP->OTLP config using the dynamic ports
+    let collector_config = generate_config("otlp", "otlp", S::name(), receiver_port, exporter_port);
+
+    let collector = CollectorProcess::start(COLLECTOR_PATH.clone(), &collector_config)
+        .await
+        .map_err(|e| format!("Failed to start collector: {}", e))?;
+
+    // Create OTLP client to send test data
+    let client_endpoint = format!("http://127.0.0.1:{}", receiver_port);
+    let client = S::connect_client(client_endpoint).await?;
+
+    // Create the test context
+    let context = TestContext {
+        client,
+        collector,
+        request_rx,
+        server_handle,
+    };
+
+    // Run the provided test logic, transferring ownership of the context
+    // The test_logic now returns the context back along with the result
+    let (mut context, result) = test_logic(context).await;
+
+    // Cleanup: drop the client connection first
+    drop(context.client);
+
+    // Send a shutdown signal to the collector process.
+    match context.collector.shutdown().await {
+        Ok(status) => {
+            if let Some(s) = status {
+                eprintln!("Collector exited with status: {}", s);
+            } else {
+                eprintln!("Collector shutdown initiated");
+            }
+        }
+        Err(e) => eprintln!("Error shutting down collector: {}", e),
+    }
+
+    drop(context.request_rx);
+
+    // Wait for the server to shut down with timeout
+    match tokio::time::timeout(
+        std::time::Duration::from_secs(SHUTDOWN_TIMEOUT_SECONDS),
+        context.server_handle,
+    )
+    .await
+    {
+        Ok(Ok(_)) => eprintln!("{} server shut down successfully", S::name()),
+        Ok(Err(e)) => eprintln!("Error shutting down {} server: {}", S::name(), e),
+        Err(_) => {
+            eprintln!("Timed out waiting for {} server to shut down", S::name());
+        }
+    }
+
+    // Return the result from the test logic
+    result
 }
